:hammer: Add Executable class
