+++ pyflow/core/kernel.py
@@ -76,9 +76,8 @@ class Kernel:
             block: CodeBlock to send the output to
             code: String representing a piece of Python code to execute
         """
-        worker = Worker(self, block, code)
-        # Change color to running
         block.run_state = ExecutableState.RUNNING
+        worker = Worker(self, block, code)
         worker.signals.stdout.connect(block.handle_stdout)
         worker.signals.image.connect(block.handle_image)
         worker.signals.finished.connect(self.run_queue)
